+++ src/Network/Network/AzureFirewallPolicy/NatRule/NewAzureFirewallPolicyNatRuleCommand.cs
@@ -16,6 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
+using System.Text.RegularExpressions;
 using Microsoft.Azure.Commands.Network.AzureFirewallPolicy;
 using Microsoft.Azure.Commands.Network.Models;
 using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
@@ -69,10 +70,16 @@ namespace Microsoft.Azure.Commands.Network
         public string[] Protocol { get; set; }
 
         [Parameter(
-            Mandatory = true,
+            Mandatory = false,
             HelpMessage = "The translated address for this NAT rule")]
         [ValidateNotNullOrEmpty]
         public string TranslatedAddress { get; set; }
+
+        [Parameter(
+            Mandatory = false,
+            HelpMessage = "The translated FQDN for this NAT rule")]
+        [ValidateNotNullOrEmpty]
+        public string TranslatedFqdn { get; set; }
 
         [Parameter(
             Mandatory = true,
@@ -84,6 +91,23 @@ namespace Microsoft.Azure.Commands.Network
         {
             base.Execute();
 
+            // Only one of TranslatedAddress or TranslatedFqdn is allowed
+            if ((TranslatedAddress != null) && (TranslatedFqdn != null))
+            {
+                throw new ArgumentException("Both TranslatedAddress and TranslatedFqdn not allowed");
+            }
+
+            // One of TranslatedAddress or TranslatedFqdn must be present
+            if ((TranslatedAddress == null) && (TranslatedFqdn == null))
+            {
+                throw new ArgumentException("Either TranslatedAddress or TranslatedFqdn is required");
+            }
+
+            if (TranslatedFqdn != null)
+            {
+                ValidateIsFqdn(TranslatedFqdn);
+            }
+
             var natRule = new PSAzureFirewallPolicyNatRule
             {
                 Name = this.Name,
@@ -93,10 +117,21 @@ namespace Microsoft.Azure.Commands.Network
                 DestinationAddresses = this.DestinationAddress?.ToList(),
                 DestinationPorts = this.DestinationPort?.ToList(),
                 TranslatedAddress = this.TranslatedAddress,
+                TranslatedFqdn = this.TranslatedFqdn,
                 TranslatedPort = this.TranslatedPort,
                 RuleType = "NatRule"
             };
             WriteObject(natRule);
+        }
+
+        private void ValidateIsFqdn(string fqdn)
+        {
+            var fqdnRegEx = new Regex("^[a-zA-Z0-9]+(([a-zA-Z0-9_\\-]*[a-zA-Z0-9]+)*\\.)*(?:[a-zA-Z0-9]{2,})$");
+
+            if (!fqdnRegEx.IsMatch(fqdn))
+            {
+                throw new ArgumentException($"Invalid value {fqdn}.");
+            }
         }
     }
 }