+++ src/Network/Network/ChangeLog.md
@@ -22,6 +22,7 @@
 * Added new cmdlets for Virtual Router Peer Routes
     - `Get-AzVirtualRouterPeerLearnedRoute`
     - `Get-AzVirtualRouterPeerAdvertisedRoute`
+* Support for Translated FQDN in NAT Rules for Azure Firewall Policy
     
 ## Version 3.5.0
 * Added Office365 Policy to VPNSite Resource